Further split constraints into basic vs compound
While `Constraints` and `AltConstraints` are indeed compound
constraints, their fundamental nature and usage within the
`Constraint` makes it advisable to keep them in base.py (avoiding
circular dependencies).
